Workaround for GitLab issue
I was originally sent this coding challenge via GitLab, and couldn't figure out a good way to do a PR in GitHub from that. This is a workaround, I'm adding the code from my GitLab repo in a ZIP file to this fork, then I'll do a PR for the fork.

Signed-off-by: ericlance0170 <38539187+ericlance0170@users.noreply.github.com>